added better authentication and authorization

+++ backend/src/services/foodService.ts
@@ -1,5 +1,8 @@
 import type { FoodCreate, FoodUpdate } from "../dtos/food/food.validation.js";
+import { Role } from "../enums/roles.js";
+import { ForbiddenError } from "../errors/ForbiddenError.js";
 import * as foodRepository from "../repositories/foodRepository.js";
+import type { SessionUser } from "../types/auth.types.js";
 import type { FoodCriteria } from "../types/food.types.js";
 import type { QueryOptions } from "../types/shared.types.js";
 
@@ -11,15 +14,37 @@ export const findByCriteria = async (criteria: FoodCriteria, options: QueryOptio
     return foodRepository.findByCriteria(criteria, options);
 };
 
-export const create = async (data: FoodCreate) => {
-    return foodRepository.create(data);
+export const create = async (data: FoodCreate, user: SessionUser) => {
+    return foodRepository.create({
+        ...data,
+        userId: user.id
+    });
 };
 
-export const update = async (id: string, data: FoodUpdate) => {
+export const update = async (id: string, data: FoodUpdate, user: SessionUser) => {
+    const food = await foodRepository.findById(id);
+
+    if (
+        food.userId !== user.id &&
+        user.role !== Role.ADMIN
+    ) {
+        throw new ForbiddenError();
+    }
+
     return foodRepository.update(id, data);
 };
 
-export const remove = async (id: string) => {
+export const remove = async (id: string, user: SessionUser) => {
+
+    const food = await foodRepository.findById(id);
+
+    if (
+        food.userId !== user.id &&
+        user.role !== Role.ADMIN
+    ) {
+        throw new ForbiddenError();
+    }
+
     await foodRepository.remove(id);
     return true;
 };
